# backend/internal/modules/user/handler.go
package user

import (
	"github.com/gofiber/fiber/v2"
)

type Handler struct {
	useCase UseCase
}

func NewHandler(useCase UseCase) *Handler {
	return &Handler{useCase: useCase}
}

// Register godoc
// @Summary      Register new user
// @Description  Register a new user account
// @Tags         User
// @Accept       json
// @Produce      json
// @Param        request body RegisterRequest true "Register Payload"
// @Success      201  {object}  RegisterResponse
// @Failure      400  {object}  map[string]interface{}
// @Failure      409  {object}  map[string]interface{}
// @Failure      500  {object}  map[string]interface{}
// @Router       /users/register [post]
func (h *Handler) Register(c *fiber.Ctx) error {
	var req RegisterRequest
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{"error": "Invalid request body"})
	}

	user, err := h.useCase.Register(c.Context(), &req)
	if err != nil {
		// 1. Cek Error Spesifik (409 Conflict)
		if err == ErrEmailTaken || err == ErrUsernameTaken {
			return c.Status(fiber.StatusConflict).JSON(fiber.Map{"error": err.Error()})
		}

		// 2. Default Error (500 Internal Server Error)
		// Tidak perlu 'if err != nil' lagi disini, karena sudah pasti error (else logic)
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Internal Server Error"})
	}

	// 3. Sukses (201 Created)
	return c.Status(fiber.StatusCreated).JSON(fiber.Map{"data": user})
}

// Login godoc
// @Summary      Login User
// @Description  Login and get JWT Token
// @Tags         User
// @Accept       json
// @Produce      json
// @Param        request body LoginRequest true "Login Payload"
// @Success      200  {object}  LoginResponse
// @Failure      400  {object}  map[string]interface{}
// @Failure      401  {object}  map[string]interface{}
// @Router       /users/login [post]
func (h *Handler) Login(c *fiber.Ctx) error {
	var req LoginRequest

	// 1. Parsing Body
	if err := c.BodyParser(&req); err != nil {
		return c.Status(fiber.StatusBadRequest).JSON(fiber.Map{
			"error": "Invalid request Body",
		})
	}

	// 2. Panggil Usecase
	resp, err := h.useCase.Login(c.Context(), &req)
	if err != nil {
		// 3. Error Handling Spesifik
		if err == ErrInvalidCredentials {
			return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{
				"error": err.Error(),
			})
		}

		// Jika Error Lain (Misal DB mati/config error), return 500
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{
			"error": "Internal Server Error",
		})
	}
	// Sukses - Return Token (Status 200 OK)
	return c.Status(fiber.StatusOK).JSON(fiber.Map{
		"data": resp,
	})
}

// GetMe godoc
// @Summary      Get Current User Profile
// @Description  Get currently logged in user profile based on JWT Token
// @Tags         User
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  UserResponse
// @Failure      401  {object}  map[string]interface{}
// @Router       /users/current [get]
func (h *Handler) GetMe(c *fiber.Ctx) error {
	// Ambil user_id dari Locals (yang diset oleh Middleware)
	userID, ok := c.Locals("user_id").(string)
	if !ok {
		return c.Status(fiber.StatusUnauthorized).JSON(fiber.Map{"error": "Unauthorized"})
	}

	resp, err := h.useCase.GetMe(c.Context(), userID)
	if err != nil {
		return c.Status(fiber.StatusInternalServerError).JSON(fiber.Map{"error": "Internal Server Error"})
	}

	return c.Status(fiber.StatusOK).JSON(fiber.Map{"data": resp})
}

func (h *Handler) RegisterRoutes(app *fiber.App, authMiddleware fiber.Handler) {
	api := app.Group("/api/users")

	api.Post("/register", h.Register)
	api.Post("/login", h.Login)

	api.Get("/current", authMiddleware, h.GetMe)
}